+++ src/entries/plugin-alert.ts
@@ -1,2 +1,2 @@
 // On-demand entry for the Alert plugin
-export { AlertPlugin } from '../plugins/AlertPlugin'
+export { alertPlugin } from '../plugins/markdown-it/alertPlugin'