+++ src/entries/plugin-katex.ts
@@ -1,2 +1,2 @@
 // On-demand entry for the KaTeX plugin
-export { katexPlugin } from '../plugins/katexPlugin'
+export { katexPlugin } from '../plugins/markdown-it/katexPlugin'