+++ src/entries/plugins/markdown-it.ts
@@ -1,4 +1,0 @@
-// Aggregate export for markdown-it built-in plugins
-// Allows import { katexPlugin } from 'mio-previewer/plugins/markdown-it'
-export { katexPlugin } from '../../plugins/katexPlugin'
-export { AlertPlugin } from '../../plugins/AlertPlugin'